requirements.txt
kafka-python==2.0.2
s3fs==2023.6.0
